87-добавление-данных-в-профилелюбом-имя-фамилия-затирает-дефолтный-аватар-смена-пароля-у-того-же-юзера-затирает-ранее-занесенные-в
поправлены замечания с гита
