+++ src/main/java/com/jm/online_store/service/impl/UserServiceImpl.java
@@ -84,7 +84,7 @@ public class UserServiceImpl implements UserService {
 
     /**
      * метод добавления нового пользователя.
-     * <p>
+     *
      * проверяется пароль на валидность, отсутствие пользователя с данным email (уникальное значение)
      *
      * @param user полученный объект User/
